фиксит детект
