+++ app/src/main/java/ru/practicum/android/diploma/data/dto/Area.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class Area(
+    @SerializedName("name") val name: String? = null,
+)
+
