+++ app/src/main/java/ru/practicum/android/diploma/data/dto/Employer.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class Employer(
+    @SerializedName("logo_urls") val logoUrls: LogoUrls? = LogoUrls(),
+    @SerializedName("name") val name: String? = null,
+)
