+++ app/src/main/java/ru/practicum/android/diploma/data/dto/Employment.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class Employment(
+    @SerializedName("name") val name: String? = null
+)
