+++ app/src/main/java/ru/practicum/android/diploma/data/dto/Experience.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.data.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class Experience(
+    @SerializedName("name") val name: String? = null
+)
