--- app/src/main/java/ru/practicum/android/diploma/data/dto/vacancySubClasses/Area.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.data.dto.vacancySubClasses
-
-import com.google.gson.annotations.SerializedName
-
-data class Area(
-    @SerializedName("name") var name: String? = null,
-)
-
