--- app/src/main/java/ru/practicum/android/diploma/data/dto/vacancySubClasses/Employer.kt
@@ -1,8 +1,0 @@
-package ru.practicum.android.diploma.data.dto.vacancySubClasses
-
-import com.google.gson.annotations.SerializedName
-
-data class Employer(
-    @SerializedName("logo_urls") var logoUrls: LogoUrls? = LogoUrls(),
-    @SerializedName("name") var name: String? = null,
-)
