--- app/src/main/java/ru/practicum/android/diploma/data/dto/vacancySubClasses/Experience.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.data.dto.vacancySubClasses
-
-import com.google.gson.annotations.SerializedName
-
-data class Experience(
-    @SerializedName("name") var name: String? = null
-)
