--- app/src/main/java/ru/practicum/android/diploma/data/dto/vacancySubClasses/LogoUrls.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.data.dto.vacancySubClasses
-
-import com.google.gson.annotations.SerializedName
-
-data class LogoUrls(
-    @SerializedName("90") var employerLogo90: String? = null,
-    @SerializedName("240") var employerLogo240: String? = null,
-    @SerializedName("original") var original: String? = null
-)
