--- app/src/main/java/ru/practicum/android/diploma/data/dto/vacancySubClasses/Schedule.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.data.dto.vacancySubClasses
-
-import com.google.gson.annotations.SerializedName
-
-data class Schedule(
-    @SerializedName("name") var name: String? = null
-)
